# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Petrobras File Transfer API
  description: |
    API RESTful para o sistema de transferencia segura de arquivos da Petrobras.
    
    ## Autenticacao
    
    A API suporta dois metodos de autenticacao:
    
    1. **Bearer Token (JWT)** - Para usuarios internos autenticados via Entra ID ou login local
    2. **Cookie de Sessao** - Alternativa ao Bearer token para aplicacoes web
    
    ## Fluxo de Compartilhamento
    
    1. Usuario interno faz upload dos arquivos
    2. Cria um compartilhamento com email do destinatario externo
    3. Supervisor aprova ou rejeita a solicitacao
    4. Destinatario recebe email com link e codigo OTP
    5. Destinatario acessa o portal, valida OTP e baixa os arquivos
    
  version: 2.0.0
  contact:
    name: Suporte Petrobras
    email: suporte@petrobras.com.br
  license:
    name: Proprietary
    url: https://petrobras.com.br
servers:
  - url: http://localhost:8000/api/v1
    description: Servidor de desenvolvimento local
  - url: https://api.petrobras.com.br/api/v1
    description: Servidor de producao
...
paths:
...
  /auth/external/request-code:
    post:
      tags: [Auth External]
      summary: Solicitar codigo OTP
      description: Envia codigo OTP para email do usuario externo
      operationId: externalRequestCode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, token]
              properties:
                email:
                  type: string
                  format: email
                token:
                  type: string
                  description: Token de acesso recebido no email
      responses:
        '200':
          description: OTP enviado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  expires_in:
                    type: integer
                    description: Segundos ate expiracao
        '429':
          description: Muitas tentativas, aguarde cooldown
...
  /files:
    get:
      tags: [Files]
      summary: Listar arquivos
      description: Lista arquivos do usuario com paginacao
      operationId: listFiles
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: area_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, deleted]
      responses:
        '200':
          description: Lista de arquivos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListResponse'
...
  /download/verify:
    post:
      tags: [Download]
      summary: Verificar token de acesso
      description: Valida token de acesso enviado por email
      operationId: verifyDownloadToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Token valido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyTokenResponse'
        '401':
          description: Token invalido ou expirado
  /download/authenticate:
    post:
      tags: [Download]
      summary: Autenticar com OTP
      description: Valida codigo OTP para acesso aos arquivos
      operationId: authenticateDownload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, otp_code]
              properties:
                token:
                  type: string
                otp_code:
                  type: string
      responses:
        '200':
          description: Autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadSessionResponse'
        '401':
          description: OTP invalido
  /download/files:
    get:
      tags: [Download]
      summary: Listar arquivos disponiveis
      description: Lista arquivos do compartilhamento autenticado
      operationId: listDownloadFiles
      security:
        - externalSession: []
      responses:
        '200':
          description: Lista de arquivos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadFilesResponse'
  /download/files/{fileId}/url:
    get:
      tags: [Download]
      summary: Obter URL de download
      description: Gera URL pre-assinada para download do arquivo
      operationId: getDownloadUrl
      security:
        - externalSession: []
      parameters:
        - $ref: '#/components/parameters/fileId'
      responses:
        '200':
          description: URL de download
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadUrlResponse'

  # ============================================================
  # AREAS - Departamentos
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para usuarios internos
    cookieAuth:
      type: apiKey
      in: cookie
      name: app_session
      description: Cookie de sessao alternativo
    externalSession:
      type: apiKey
      in: cookie
      name: external_session
      description: Cookie de sessao para usuarios externos
  parameters:
...
    fileId:
      name: fileId
      in: path
      required: true
      schema:
        type: integer
      description: ID do arquivo
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        type:
          type: string
          enum: [externo, internal, supervisor]
        status:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    File:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        mime_type:
          type: string
        size_bytes:
          type: integer
        size_formatted:
          type: string
        area_id:
          type: integer
        uploaded_by:
          $ref: '#/components/schemas/User'
        status:
          type: boolean
        created_at:
          type: string
          format: date-time
    FileListResponse:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/File'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    VerifyTokenResponse:
      type: object
      properties:
        valid:
          type: boolean
        share_id:
          type: integer
        recipient_email:
          type: string
        expires_at:
          type: string
          format: date-time
        otp_required:
          type: boolean
    DownloadSessionResponse:
      type: object
      properties:
        session_token:
          type: string
        expires_at:
          type: string
          format: date-time
        share:
          type: object
          properties:
            id:
              type: integer
            sender_name:
              type: string
            files_count:
              type: integer
    DownloadFilesResponse:
      type: object
      properties:
        share:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            sender:
              type: string
            expires_at:
              type: string
              format: date-time
        files:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              size:
                type: string
              mime_type:
                type: string
              downloaded:
                type: boolean
    DownloadUrlResponse:
      type: object
      properties:
        url:
          type: string
          format: uri
        expires_in:
          type: integer
        filename:
          type: string

    # Area Schemas
...
    Pagination:
      type: object
      properties:
        current_page:
          type: integer
        total_pages:
          type: integer
        total_items:
          type: integer
        items_per_page:
          type: integer